implement typing capabilities in KeyboardArea
